feat(react-query): add getStatusTable function
